Clean PR: removed prints, unrelated changes, move layer_type helper to utils, fix softcap handling and density logging
