+++ sparse_attention_hub/sparse_attention/research_attention/base.py
@@ -122,12 +122,10 @@ class ResearchAttention(SparseAttention):
             )
 
         layer_type: str = str(kwargs.get("layer_type", "unknown"))
-        print("DEBUG:", kwargs.get("layer_idx"), kwargs.get("layer_type"))
         assert layer_type in ["full_attention", "sliding_attention", "unknown"]
         if MicroMetricLogger().is_metric_enabled("research_attention_density"):
             # Density is only meaningful for dense/full-attention layers.
             if layer_type == "full_attention":
-                print("LOGGING DENSITY FOR LAYER:", kwargs.get("layer_idx"))
                 MicroMetricLogger().log(
                     "research_attention_density",
                     sparse_attention_mask.get_density(),
